Adjust the namespace.
